:sparkles: feat: 업무 일지 저장 및 ID 가져오기 로직 추가
Fixes: #19

+++ app/src/main/java/com/schopenhauer/nous/data/local/model/TaskEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "tasks")
 data class TaskEntity (
-	@PrimaryKey(autoGenerate = true) val id: Int = 0,
-	val journalId: Int = 0,
+	@PrimaryKey(autoGenerate = true) val id: Long = 0,
+	val journalId: Long = 0,
 	val content: String,
 )
