+++ app/src/main/java/com/schopenhauer/nous/domain/model/Task.kt
@@ -1,7 +1,7 @@
 package com.schopenhauer.nous.domain.model
 
 data class Task(
-	val id: Int,
-	val journalId: Int = 0,
+	val id: Long,
+	val journalId: Long = 0,
 	val content: String
 )
