+++ app/src/main/java/com/schopenhauer/nous/ui/journal/list/JournalAdapter.kt
@@ -9,7 +9,7 @@ import com.schopenhauer.nous.databinding.ListItemJournalBinding
 import com.schopenhauer.nous.domain.model.Journal
 
 class JournalAdapter(
-	private val itemClickListener: (Int) -> Unit
+	private val itemClickListener: (Long) -> Unit
 ) : ListAdapter<Journal, JournalAdapter.JournalViewHolder>(diffUtil) {
 
 	override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): JournalViewHolder {
